+++ src/lib/uart.h
@@ -6,7 +6,7 @@
  * Cabecera para el manejo de la UART.
  **/
 
-void configUart(uint16_t baud);
+void configUart(uint32_t baud);
 void setUartTx(uint8_t data);
 uint8_t getUartData(void);
 void setUartRxIsr(void (*isr)(uint8_t data));
